// src/practices/common/models/location.model.ts
import { Room } from './room.model';

export interface Location {
    LocationId:                          number;
    NameLine1:                           string;
    NameLine2:                           null;
    NameAbbreviation:                    string;
    ImageFile:                           null;
    LogoFile:                            null;
    Website:                             null;
    Timezone:                            string;
    DeactivationTimeUtc:                 null;
    AddressLine1:                        string;
    AddressLine2:                        null;
    City:                                string;
    State:                               string;
    ZipCode:                             string;
    Email:                               string;
    PrimaryPhone:                        null;
    SecondaryPhone:                      null;
    Fax:                                 null;
    TaxId:                               null;
    TypeTwoNpi:                          null;
    TaxonomyId:                          null;
    LicenseNumber:                       null;
    ProviderTaxRate:                     null;
    SalesAndUseTaxRate:                  null;
    DefaultFinanceCharge:                null;
    AccountsOverDue:                     null;
    MinimumFinanceCharge:                null;
    Rooms:                               Room[];
    AdditionalIdentifiers:               any[];
    MasterLocationAdditionalIdentifiers: any[];
    MerchantId:                          null;
    IsPaymentGatewayEnabled:             boolean;
    DisplayCardsOnEstatement:            boolean;
    AcceptMasterCardOnEstatement:        boolean;
    AcceptDiscoverOnEstatement:          boolean;
    AcceptVisaOnEstatement:              boolean;
    AcceptAmericanExpressOnEstatement:   boolean;
    IncludeCvvCodeOnEstatement:          boolean;
    RemitAddressSource:                  number;
    RemitOtherLocationId:                null;
    RemitToNameLine1:                    null;
    RemitToNameLine2:                    null;
    RemitToAddressLine1:                 null;
    RemitToAddressLine2:                 null;
    RemitToCity:                         null;
    RemitToState:                        null;
    RemitToZipCode:                      null;
    RemitToPrimaryPhone:                 null;
    InsuranceRemittanceAddressSource:    number;
    InsuranceRemittanceOtherLocationId:  null;
    InsuranceRemittanceNameLine1:        null;
    InsuranceRemittanceNameLine2:        null;
    InsuranceRemittanceAddressLine1:     null;
    InsuranceRemittanceAddressLine2:     null;
    InsuranceRemittanceCity:             null;
    InsuranceRemittanceState:            null;
    InsuranceRemittanceZipCode:          null;
    InsuranceRemittancePrimaryPhone:     null;
    InsuranceRemittanceTaxId:            null;
    InsuranceRemittanceTypeTwoNpi:       null;
    InsuranceRemittanceLicenseNumber:    null;
    FeeListId:                           number;
    IsRxRegistered:                      boolean;
    DataTag:                             string;
    UserModified:                        string;
    DateModified:                        string;
    PlaceOfTreatment:                    number;
}